Resolve ClassVar inside deferred annotations.
